Whitespace cleanup
This turns tabs to spaces for files that are part of the
TitleWait code, and canonizes on LF only line endings.  It
also removes stray whitespace at ends of lines and ensures
each file ends in a newline.

(NTProcessInfo is considered a third party file for now,
so it is left with its tabs)