Update stylization to my 2022 tastes
Since 2008 I've changed how I space things a bit, making use
of half-indents for case labels or public/private/protected
labels...and generally spacing comments out more from the
code they describe.

This makes those changes, as well as a tweak to remove the
need for an indentation level in the DebugLoop code.