Fix compiler warning on wide strings
In order to call the system() function, the wide string of
the command line arguments had to be converted into a non-wide
string.  The method of doing so was proposed by a StackOverflow
answer which just constructed a std::string from the .begin()
and .end() of the wide string.  This should have triggered
a compiler warning, but did not.

Ultimately MSVC started issuing the warning.  Some suggestions
for how to use C++11 `<codecvt>` functions and std::transform
were made, but these wound up being deprecated in C++17.

A Windows-only method was proposed as a workaround here:

   https://stackoverflow.com/a/69410299